requirements.txt
tensorflow
matplotlib
numpy
flask
pillow